Expose properties so that an app can determine if AOT is active
This commit introduces an internal optimization which exposes new
properties to the application runtime, so that it can determine if
AOT is enabled. An easy way is for an application to define a bean
annotated with `@ConfigurationProperties`:

```java
@ConfigurationProperties("micronaut.aot")
public class AotConfiguration {
    private boolean enabled;
    private String runtime;
    private List<String> optimizations;
    ...
}
```

The diagnostics that we expose are intentionally limited.
For example, we don't want to expose the whole configuration,
as it would contain entries which are sensitive to the build
environment (absolute paths for example) which would defeat
caching.
