+++ aot-core/src/test/groovy/io/micronaut/aot/core/codegen/ApplicationContextConfigurerGeneratorTest.groovy
@@ -41,10 +41,28 @@ class ApplicationContextConfigurerGeneratorTest extends AbstractSourceGeneratorS
             hasClass('AOTApplicationContextConfigurer') {
                 withSources '''package io.micronaut.test;
 
+import io.micronaut.context.ApplicationContextBuilder;
 import io.micronaut.context.ApplicationContextConfigurer;
+import java.lang.Override;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
 
 public class AOTApplicationContextConfigurer implements ApplicationContextConfigurer {
   static {
+  }
+
+  private static List list0() {
+    return Collections.emptyList();
+  }
+
+  @Override
+  public void configure(ApplicationContextBuilder builder) {
+    builder.properties(new HashMap() {{
+        put("micronaut.aot.enabled", true);
+        put("micronaut.aot.runtime", "JIT");
+        put("micronaut.aot.optimizations", list0());
+        }});
   }
 }'''
             }
@@ -65,7 +83,12 @@ public class AOTApplicationContextConfigurer implements ApplicationContextConfig
             hasClass('AOTApplicationContextConfigurer') {
                 withSources '''package io.micronaut.test;
 
+import io.micronaut.context.ApplicationContextBuilder;
 import io.micronaut.context.ApplicationContextConfigurer;
+import java.lang.Override;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
 
 public class AOTApplicationContextConfigurer implements ApplicationContextConfigurer {
   static {
@@ -77,6 +100,22 @@ public class AOTApplicationContextConfigurer implements ApplicationContextConfig
   }
 
   private static void initializer2() {
+  }
+
+  private static List list0() {
+    List result = new ArrayList<>(2);
+    result.add("static-init");
+    result.add("static-init");
+    return result;
+  }
+
+  @Override
+  public void configure(ApplicationContextBuilder builder) {
+    builder.properties(new HashMap() {{
+        put("micronaut.aot.enabled", true);
+        put("micronaut.aot.runtime", "JIT");
+        put("micronaut.aot.optimizations", list0());
+        }});
   }
 }
 '''
